Added endpoints

+++ app/models/product.py
@@ -7,6 +7,6 @@ class Product(Base):
     __tablename__ = "product"
     p_id = Column(Integer, primary_key=True, index=True, autoincrement=True)
     name = Column(String(255), nullable=False)
-    price = Column(Integer)
+    price = Column(Float)
     w_id = Column(Integer, ForeignKey('weed.w_id'), nullable=False)
